+++ _is_piped.d.ts
@@ -1,0 +1,2 @@
+export declare function _is_piped(): boolean;
+export { _is_piped as _is__piped };