+++ _piped_a1.d.ts
@@ -1,0 +1,2 @@
+export declare function _piped_a1(): Promise<string[]>;
+export { _piped_a1 as _a1__piped };