+++ _piped_txt.d.ts
@@ -1,0 +1,2 @@
+export declare function _piped_txt(): Promise<string>;
+export { _piped_txt as _txt__piped };