+++ index.ts
@@ -1,1 +1,3 @@
-export * from './lib'
+export * from './_is_piped'
+export * from './_piped_a1'
+export * from './_piped_txt'